refactor: Extract AwsProfileCredential
